+++ src/api/admin/question-manage/admin-search-by-id.ts
@@ -1,4 +1,4 @@
-import { server_axiosInstance } from '../../utils/axios';
+import { server_axiosInstance } from '../../../utils/axios';
 
 export async function SearchById(id: number) {
   try {
